requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.20
asyncio==3.4.3
psycopg2-binary
